+++ src/utils.h
@@ -56,7 +56,7 @@ dyn_str *dyn_str_from_string(const char *str);
 
 void exit_error(const char *format, ...); 
 size_t num_len(const int n);
-const char *num_to_str(const int n);
+const char *num_to_str(const ssize_t n);
 
 void set_timer(struct timespec *timer);
 double get_ms_elapsed(const struct timespec *start);
